refined test data
